refactor(#3038): VoiceBargeInRuntime S2 — extract progress playback cluster
S2 scope: extract the next cohesive VoiceBargeInRuntime cluster from the S1 concern map without behavior changes. The S1 map shows playback is the next largest aggregate after routing/dispatch, but playback is explicitly split across progress playback, live-cut, and final-result playback. This slice picks the progress playback subcluster because it moves cleanly: progress broadcast/pub-sub, progress worker state handling, text mirroring, progress summaries, idle notices, processing chime pathing, and one-shot progress/chime/ack audio playback.

Cluster boundary: moved `subscribe_progress`, `publish_progress`, `publish_progress_for_playback`, `spawn_progress_worker`, `handle_progress_event`, `flush_due_progress_summaries`, `emit_due_idle_notices`, `mirror_progress_line`, `speak_progress_summary`, `speak_progress_text`, `play_processing_chime`, `processing_chime_path`, `play_acknowledgement`, and `play_progress_audio` verbatim into `src/services/discord/voice_barge_in/progress_playback.rs` as an inherent impl child module. `speak_voice_background_completion_summary` and `spawn_spoken_result_playback` stay in the root as final-result playback, while `reset_after_playback_start_with_owner`, `clear_playback_if_owner`, `observe_live_pcm_i16`, and monitor sensitivity helpers stay in the root as the live-cut subcluster.

No await point, lock scope, spawn, map-entry, cancellation, or playback ordering changed. The only moved-method signature edits are visibility-only: `speak_progress_text`, `play_processing_chime`, and `play_acknowledgement` are `pub(super)` because parent routing/final-result code still calls them. No struct fields were widened. Existing `#[cfg(test)]` blocks inside moved methods are preserved exactly.

LoC accounting: `voice_barge_in.rs` production LoC drops 4350 -> 3932 (-418). New child `voice_barge_in/progress_playback.rs` is 423 production LoC, below the 1000-line giant threshold, so only the root ratchet is lowered.

Token proof: normalized method-token comparison against HEAD is identical after visibility-only normalization. Per moved method:

| Method | Method tokens | Body tokens | Normalized SHA |
| --- | ---: | ---: | --- |
| `subscribe_progress` | 32 -> 32 | 7 -> 7 | `4d633f0f24d6` |
| `publish_progress` | 41 -> 41 | 11 -> 11 | `b73462491438` |
| `publish_progress_for_playback` | 100 -> 100 | 63 -> 63 | `74d7e90d34d5` |
| `spawn_progress_worker` | 252 -> 252 | 215 -> 215 | `9af55bdde36d` |
| `handle_progress_event` | 371 -> 371 | 333 -> 333 | `ddd22f886361` |
| `flush_due_progress_summaries` | 198 -> 198 | 164 -> 164 | `e2dda739ba3d` |
| `emit_due_idle_notices` | 321 -> 321 | 287 -> 287 | `7c2f23f9ef09` |
| `mirror_progress_line` | 168 -> 168 | 141 -> 141 | `979133b328e3` |
| `speak_progress_summary` | 77 -> 77 | 43 -> 43 | `82b4e76f04de` |
| `speak_progress_text` | 77 -> 77 | 39 -> 39 | `606b6e2c5b0b` |
| `play_processing_chime` | 63 -> 63 | 36 -> 36 | `2b3a86078f78` |
| `processing_chime_path` | 156 -> 156 | 142 -> 142 | `01016e362f63` |
| `play_acknowledgement` | 50 -> 50 | 19 -> 19 | `42c714c8f4d4` |
| `play_progress_audio` | 527 -> 527 | 490 -> 490 | `ed9ca23d398f` |
| **Total** | **2433 -> 2433** | **1990 -> 1990** | |

Gates run before commit:

- cargo fmt --check
- cargo check --lib
- cargo clippy --lib -- -D warnings
- cargo test --lib voice && cargo test --lib discord 2>&1 | tail -1
- python3 scripts/audit_maintainability.py --check
